Separate mavlink functionality from UI
