+++ qgroundcontrol.pro
@@ -369,6 +369,7 @@ INCLUDEPATH += \
     src/Geo \
     src/GPS \
     src/Joystick \
+    src/Mavlink \
     src/PlanView \
     src/MissionManager \
     src/PositionManager \
@@ -381,7 +382,7 @@ INCLUDEPATH += \
     src/Vehicle/Actuators \
     src/Vehicle/Components \
     src/Vehicle/FactGroups \
-    src/Vehicle/LibEvents \
+    src/Mavlink/LibEvents \
     src/Audio \
     src/comm \
     src/comm/MockLink \
@@ -430,7 +431,10 @@ HEADERS += \
     src/AnalyzeView/LogDownloadController.h \
     src/AnalyzeView/PX4LogParser.h \
     src/AnalyzeView/ULogParser.h \
-    src/AnalyzeView/MavlinkConsoleController.h \
+    src/AnalyzeView/MAVLinkChartController.h \
+    src/AnalyzeView/MAVLinkConsoleController.h \
+    src/AnalyzeView/MAVLinkMessage.h \
+    src/AnalyzeView/MAVLinkSystem.h \
     src/Audio/AudioOutput.h \
     src/Vehicle/Autotune.h \
     src/Camera/MavlinkCameraControl.h \
@@ -561,9 +565,9 @@ HEADERS += \
     src/Vehicle/Actuators/GeometryImage.h \
     src/Vehicle/Actuators/Mixer.h \
     src/Vehicle/Actuators/MotorAssignment.h \
-    src/Vehicle/LibEvents/EventHandler.h \
-    src/Vehicle/LibEvents/HealthAndArmingCheckReport.h \
-    src/Vehicle/LibEvents/libevents_includes.h \
+    src/Mavlink/LibEvents/EventHandler.h \
+    src/Mavlink/LibEvents/HealthAndArmingCheckReport.h \
+    src/Mavlink/LibEvents/libevents_includes.h \
     src/Vehicle/Components/CompInfo.h \
     src/Vehicle/Components/CompInfoActuators.h \
     src/Vehicle/Components/CompInfoEvents.h \
@@ -574,15 +578,15 @@ HEADERS += \
     src/Vehicle/Components/ComponentInformationTranslation.h \
     src/Vehicle/FTPManager.h \
     src/Vehicle/FactGroups/GPSRTKFactGroup.h \
-    src/Vehicle/ImageProtocolManager.h \
+    src/Mavlink/ImageProtocolManager.h \
     src/Vehicle/InitialConnectStateMachine.h \
     src/Vehicle/MAVLinkLogManager.h \
-    src/Vehicle/MAVLinkStreamConfig.h \
+    src/Mavlink/MAVLinkStreamConfig.h \
     src/Vehicle/MultiVehicleManager.h \
     src/Vehicle/RemoteIDManager.h \
     src/Utilities/StateMachine.h \
     src/Vehicle/StandardModes.h \
-    src/Vehicle/SysStatusSensorInfo.h \
+    src/Mavlink/SysStatusSensorInfo.h \
     src/Vehicle/FactGroups/TerrainFactGroup.h \
     src/Vehicle/TerrainProtocolHandler.h \
     src/Vehicle/TrajectoryPoints.h \
@@ -610,12 +614,15 @@ HEADERS += \
     src/comm/LinkManager.h \
     src/comm/LogReplayLink.h \
     src/comm/MAVLinkProtocol.h \
-    src/comm/QGCMAVLink.h \
+    src/MAVLink/QGCMAVLink.h \
+    src/MAVLink/MAVLinkFTP.h \
+    src/MAVLink/MAVLinkLib.h \
     src/comm/TCPLink.h \
     src/comm/UDPLink.h \
     src/comm/UdpIODevice.h \
     src/Vehicle/UASMessageHandler.h \
     src/AnalyzeView/GeoTagController.h \
+    src/AnalyzeView/GeoTagWorker.h \
     src/AnalyzeView/ExifParser.h \
     src/Viewer3D/CityMapGeometry.h \
     src/Viewer3D/OsmParser.h \
@@ -698,7 +705,10 @@ SOURCES += \
     src/AnalyzeView/LogDownloadController.cc \
     src/AnalyzeView/PX4LogParser.cc \
     src/AnalyzeView/ULogParser.cc \
-    src/AnalyzeView/MavlinkConsoleController.cc \
+    src/AnalyzeView/MAVLinkChartController.cc \
+    src/AnalyzeView/MAVLinkConsoleController.cc \
+    src/AnalyzeView/MAVLinkMessage.cc \
+    src/AnalyzeView/MAVLinkSystem.cc \
     src/Audio/AudioOutput.cc \
     src/Vehicle/Autotune.cpp \
     src/Camera/MavlinkCameraControl.cc \
@@ -825,9 +835,9 @@ SOURCES += \
     src/Vehicle/Actuators/GeometryImage.cc \
     src/Vehicle/Actuators/Mixer.cc \
     src/Vehicle/Actuators/MotorAssignment.cc \
-    src/Vehicle/LibEvents/EventHandler.cc \
-    src/Vehicle/LibEvents/HealthAndArmingCheckReport.cc \
-    src/Vehicle/LibEvents/logging.cpp \
+    src/Mavlink/LibEvents/EventHandler.cc \
+    src/Mavlink/LibEvents/HealthAndArmingCheckReport.cc \
+    src/Mavlink/LibEvents/logging.cpp \
     src/Vehicle/Components/CompInfo.cc \
     src/Vehicle/Components/CompInfoActuators.cc \
     src/Vehicle/Components/CompInfoEvents.cc \
@@ -838,15 +848,15 @@ SOURCES += \
     src/Vehicle/Components/ComponentInformationTranslation.cc \
     src/Vehicle/FTPManager.cc \
     src/Vehicle/FactGroups/GPSRTKFactGroup.cc \
-    src/Vehicle/ImageProtocolManager.cc \
+    src/Mavlink/ImageProtocolManager.cc \
     src/Vehicle/InitialConnectStateMachine.cc \
     src/Vehicle/MAVLinkLogManager.cc \
-    src/Vehicle/MAVLinkStreamConfig.cc \
+    src/Mavlink/MAVLinkStreamConfig.cc \
     src/Vehicle/MultiVehicleManager.cc \
     src/Vehicle/RemoteIDManager.cc \
     src/Utilities/StateMachine.cc \
     src/Vehicle/StandardModes.cc \
-    src/Vehicle/SysStatusSensorInfo.cc \
+    src/Mavlink/SysStatusSensorInfo.cc \
     src/Vehicle/FactGroups/TerrainFactGroup.cc \
     src/Vehicle/TerrainProtocolHandler.cc \
     src/Vehicle/TrajectoryPoints.cc \
@@ -874,13 +884,15 @@ SOURCES += \
     src/comm/LinkManager.cc \
     src/comm/LogReplayLink.cc \
     src/comm/MAVLinkProtocol.cc \
-    src/comm/QGCMAVLink.cc \
+    src/MAVLink/QGCMAVLink.cc \
+    src/MAVLink/MAVLinkFTP.cc \
     src/comm/TCPLink.cc \
     src/comm/UDPLink.cc \
     src/comm/UdpIODevice.cc \
     src/main.cc \
     src/Vehicle/UASMessageHandler.cc \
     src/AnalyzeView/GeoTagController.cc \
+    src/AnalyzeView/GeoTagWorker.cc \
     src/AnalyzeView/ExifParser.cc \
     src/Viewer3D/CityMapGeometry.cc \
     src/Viewer3D/OsmParser.cc \
