+++ src/AnalyzeView/LogDownloadController.cc
@@ -17,6 +17,7 @@
 #include "FactSystem.h"
 #include "Vehicle.h"
 #include "SettingsManager.h"
+#include "MAVLinkProtocol.h"
 #include "QGCLoggingCategory.h"
 
 #include <QtCore/QBitArray>
