+++ src/AnalyzeView/MAVLinkConsoleController.cc
@@ -7,24 +7,25 @@
  *
  ****************************************************************************/
 
-#include "MavlinkConsoleController.h"
+#include "MAVLinkConsoleController.h"
 #include "QGCApplication.h"
 #include "QGCToolbox.h"
 #include "MultiVehicleManager.h"
 #include "Vehicle.h"
+#include "MAVLinkProtocol.h"
 
 #include <QtWidgets/QApplication>
 #include <QtGui/QClipboard>
 
-MavlinkConsoleController::MavlinkConsoleController()
+MAVLinkConsoleController::MAVLinkConsoleController()
     : QStringListModel()
 {
     auto *manager = qgcApp()->toolbox()->multiVehicleManager();
-    connect(manager, &MultiVehicleManager::activeVehicleChanged, this, &MavlinkConsoleController::_setActiveVehicle);
+    connect(manager, &MultiVehicleManager::activeVehicleChanged, this, &MAVLinkConsoleController::_setActiveVehicle);
     _setActiveVehicle(manager->activeVehicle());
 }
 
-MavlinkConsoleController::~MavlinkConsoleController()
+MAVLinkConsoleController::~MAVLinkConsoleController()
 {
     if (_vehicle) {
         QByteArray msg;
@@ -33,7 +34,7 @@ MavlinkConsoleController::~MavlinkConsoleController()
 }
 
 void
-MavlinkConsoleController::sendCommand(QString command)
+MAVLinkConsoleController::sendCommand(QString command)
 {
     // there might be multiple commands, add them separately to the history
     QStringList lines = command.split('\n');
@@ -48,19 +49,19 @@ MavlinkConsoleController::sendCommand(QString command)
 }
 
 QString
-MavlinkConsoleController::historyUp(const QString& current)
+MAVLinkConsoleController::historyUp(const QString& current)
 {
     return _history.up(current);
 }
 
 QString
-MavlinkConsoleController::historyDown(const QString& current)
+MAVLinkConsoleController::historyDown(const QString& current)
 {
     return _history.down(current);
 }
 
 QString
-MavlinkConsoleController::handleClipboard(const QString& command_pre)
+MAVLinkConsoleController::handleClipboard(const QString& command_pre)
 {
     QString clipboardData = command_pre + QApplication::clipboard()->text();
     int lastLinePos = clipboardData.lastIndexOf('\n');
@@ -73,7 +74,7 @@ MavlinkConsoleController::handleClipboard(const QString& command_pre)
 }
 
 void
-MavlinkConsoleController::_setActiveVehicle(Vehicle* vehicle)
+MAVLinkConsoleController::_setActiveVehicle(Vehicle* vehicle)
 {
     for (auto &con : _uas_connections)
         disconnect(con);
@@ -88,12 +89,12 @@ MavlinkConsoleController::_setActiveVehicle(Vehicle* vehicle)
         _cursorY = 0;
         _cursorX = 0;
         _cursor_home_pos = -1;
-        _uas_connections << connect(_vehicle, &Vehicle::mavlinkSerialControl, this, &MavlinkConsoleController::_receiveData);
+        _uas_connections << connect(_vehicle, &Vehicle::mavlinkSerialControl, this, &MAVLinkConsoleController::_receiveData);
     }
 }
 
 void
-MavlinkConsoleController::_receiveData(uint8_t device, uint8_t, uint16_t, uint32_t, QByteArray data)
+MAVLinkConsoleController::_receiveData(uint8_t device, uint8_t, uint16_t, uint32_t, QByteArray data)
 {
     if (device != SERIAL_CONTROL_DEV_SHELL)
         return;
@@ -131,7 +132,7 @@ MavlinkConsoleController::_receiveData(uint8_t device, uint8_t, uint16_t, uint32
 }
 
 void
-MavlinkConsoleController::_sendSerialData(QByteArray data, bool close)
+MAVLinkConsoleController::_sendSerialData(QByteArray data, bool close)
 {
     if (!_vehicle) {
         qWarning() << "Internal error";
@@ -172,7 +173,7 @@ MavlinkConsoleController::_sendSerialData(QByteArray data, bool close)
 }
 
 bool
-MavlinkConsoleController::_processANSItext(QByteArray &line)
+MAVLinkConsoleController::_processANSItext(QByteArray &line)
 {
     // Iterate over the incoming buffer to parse off known ANSI control codes
     for (int i = 0; i < line.size(); i++) {
@@ -236,7 +237,7 @@ MavlinkConsoleController::_processANSItext(QByteArray &line)
 }
 
 QString
-MavlinkConsoleController::transformLineForRichText(const QString& line) const
+MAVLinkConsoleController::transformLineForRichText(const QString& line) const
 {
     QString ret = line.toHtmlEscaped().replace(" ","&nbsp;").replace("\t", "&nbsp;&nbsp;&nbsp;&nbsp;");
 
@@ -250,7 +251,7 @@ MavlinkConsoleController::transformLineForRichText(const QString& line) const
 }
 
 QString
-MavlinkConsoleController::getText() const
+MAVLinkConsoleController::getText() const
 {
     QString ret;
     if (rowCount() > 0) {
@@ -264,7 +265,7 @@ MavlinkConsoleController::getText() const
 }
 
 void
-MavlinkConsoleController::writeLine(int line, const QByteArray &text)
+MAVLinkConsoleController::writeLine(int line, const QByteArray &text)
 {
     auto rc = rowCount();
     if (line >= rc) {
@@ -286,7 +287,7 @@ MavlinkConsoleController::writeLine(int line, const QByteArray &text)
     _cursorX += text.size();
 }
 
-void MavlinkConsoleController::CommandHistory::append(const QString& command)
+void MAVLinkConsoleController::CommandHistory::append(const QString& command)
 {
     if (command.length() > 0) {
 
@@ -302,7 +303,7 @@ void MavlinkConsoleController::CommandHistory::append(const QString& command)
     _index = _history.length();
 }
 
-QString MavlinkConsoleController::CommandHistory::up(const QString& current)
+QString MAVLinkConsoleController::CommandHistory::up(const QString& current)
 {
     if (_index <= 0)
         return current;
@@ -314,7 +315,7 @@ QString MavlinkConsoleController::CommandHistory::up(const QString& current)
     return "";
 }
 
-QString MavlinkConsoleController::CommandHistory::down(const QString& current)
+QString MAVLinkConsoleController::CommandHistory::down(const QString& current)
 {
     if (_index >= _history.length())
         return current;
