+++ src/AnalyzeView/MAVLinkInspectorController.h
@@ -14,238 +14,18 @@
 #pragma once
 
 #include "QGCMAVLink.h"
+#include "MAVLinkChartController.h"
+#include "MAVLinkSystem.h"
 #include "QmlObjectListModel.h"
 #include <QtCore/QObject>
 #include <QtCore/QString>
 #include <QtCore/QTimer>
-#include <QtCore/QDateTime>
-#include <QtCore/QVariantList>
 #include <QtCore/QLoggingCategory>
-#include <QtCharts/QAbstractSeries>
 
 Q_DECLARE_LOGGING_CATEGORY(MAVLinkInspectorLog)
 
-class QGCMAVLinkMessage;
-class QGCMAVLinkSystem;
-class MAVLinkChartController;
-class MAVLinkInspectorController;
 class Vehicle;
 class LinkInterface;
-
-//-----------------------------------------------------------------------------
-/// MAVLink message field
-class QGCMAVLinkMessageField : public QObject {
-    Q_OBJECT
-public:
-    Q_PROPERTY(QString          name        READ name       CONSTANT)
-    Q_PROPERTY(QString          label       READ label      CONSTANT)
-    Q_PROPERTY(QString          type        READ type       CONSTANT)
-    Q_PROPERTY(QString          value       READ value      NOTIFY valueChanged)
-    Q_PROPERTY(bool             selectable  READ selectable NOTIFY selectableChanged)
-    Q_PROPERTY(int              chartIndex  READ chartIndex CONSTANT)
-    Q_PROPERTY(QAbstractSeries* series      READ series     NOTIFY seriesChanged)
-
-    QGCMAVLinkMessageField(QGCMAVLinkMessage* parent, QString name, QString type);
-
-    QString         name            () { return _name;  }
-    QString         label           ();
-    QString         type            () { return _type;  }
-    QString         value           () { return _value; }
-    bool            selectable      () const{ return _selectable; }
-    bool            selected        () { return _pSeries != nullptr; }
-    QAbstractSeries*series          () { return _pSeries; }
-    QList<QPointF>* values          () { return &_values;}
-    qreal           rangeMin        () const{ return _rangeMin; }
-    qreal           rangeMax        () const{ return _rangeMax; }
-    int             chartIndex      ();
-
-    void            setSelectable   (bool sel);
-    void            updateValue     (QString newValue, qreal v);
-
-    void            addSeries       (MAVLinkChartController* chart, QAbstractSeries* series);
-    void            delSeries       ();
-    void            updateSeries    ();
-
-signals:
-    void            seriesChanged       ();
-    void            selectableChanged   ();
-    void            valueChanged        ();
-
-private:
-    QString     _type;
-    QString     _name;
-    QString     _value;
-    bool        _selectable = true;
-    int         _dataIndex  = 0;
-    qreal       _rangeMin   = 0;
-    qreal       _rangeMax   = 0;
-
-    QAbstractSeries*    _pSeries = nullptr;
-    QGCMAVLinkMessage*  _msg     = nullptr;
-    MAVLinkChartController*      _chart   = nullptr;
-    QList<QPointF>      _values;
-};
-
-//-----------------------------------------------------------------------------
-/// MAVLink message
-class QGCMAVLinkMessage : public QObject {
-    Q_OBJECT
-public:
-    Q_PROPERTY(quint32              id              READ id             CONSTANT)
-    Q_PROPERTY(quint32              sysId           READ sysId          CONSTANT)
-    Q_PROPERTY(quint32              compId          READ compId         CONSTANT)
-    Q_PROPERTY(QString              name            READ name           CONSTANT)
-    Q_PROPERTY(qreal                actualRateHz    READ actualRateHz   NOTIFY actualRateHzChanged)
-    Q_PROPERTY(int32_t              targetRateHz    READ targetRateHz   NOTIFY targetRateHzChanged)
-    Q_PROPERTY(quint64              count           READ count          NOTIFY countChanged)
-    Q_PROPERTY(QmlObjectListModel*  fields          READ fields         CONSTANT)
-    Q_PROPERTY(bool                 fieldSelected   READ fieldSelected  NOTIFY fieldSelectedChanged)
-    Q_PROPERTY(bool                 selected        READ selected       NOTIFY selectedChanged)
-
-    QGCMAVLinkMessage   (QObject* parent, mavlink_message_t* message);
-    ~QGCMAVLinkMessage  ();
-
-    quint32             id              () const { return _message.msgid;  }
-    quint8              sysId           () const { return _message.sysid; }
-    quint8              compId          () const { return _message.compid; }
-    QString             name            () const { return _name;  }
-    qreal               actualRateHz    () const { return _actualRateHz; }
-    int32_t             targetRateHz    () const { return _targetRateHz; }
-    quint64             count           () const { return _count; }
-    quint64             lastCount       () const { return _lastCount; }
-    QmlObjectListModel* fields          () { return &_fields; }
-    bool                fieldSelected   () const { return _fieldSelected; }
-    bool                selected        () const { return _selected; }
-
-    void                updateFieldSelection();
-    void                update          (mavlink_message_t* message);
-    void                updateFreq      ();
-    void                setSelected     (bool sel);
-    void                setTargetRateHz (int32_t rate);
-
-signals:
-    void countChanged();
-    void actualRateHzChanged();
-    void targetRateHzChanged();
-    void fieldSelectedChanged();
-    void selectedChanged();
-
-private:
-    void _updateFields(void);
-
-    QmlObjectListModel  _fields;
-    QString             _name;
-    qreal               _actualRateHz   = 0.0;
-    int32_t             _targetRateHz   = 0;
-    uint64_t            _count          = 1;
-    uint64_t            _lastCount      = 0;
-    mavlink_message_t   _message;
-    bool                _fieldSelected  = false;
-    bool                _selected       = false;
-};
-
-//-----------------------------------------------------------------------------
-/// Vehicle MAVLink message belongs to
-class QGCMAVLinkSystem : public QObject {
-    Q_OBJECT
-public:
-    Q_PROPERTY(quint8               id              READ id                                 CONSTANT)
-    Q_PROPERTY(QmlObjectListModel*  messages        READ messages                           CONSTANT)
-    Q_PROPERTY(QList<int>           compIDs         READ compIDs                            NOTIFY compIDsChanged)
-    Q_PROPERTY(QStringList          compIDsStr      READ compIDsStr                         NOTIFY compIDsChanged)
-    Q_PROPERTY(int                  selected        READ selected       WRITE setSelected   NOTIFY selectedChanged)
-
-    QGCMAVLinkSystem   (QObject* parent, quint8 id);
-    ~QGCMAVLinkSystem  ();
-
-    quint8              id              () const{ return _id; }
-    QmlObjectListModel* messages        () { return &_messages; }
-    QList<int>          compIDs         () { return _compIDs; }
-    QStringList         compIDsStr      () { return _compIDsStr; }
-    int                 selected        () const{ return _selected; }
-
-    void                setSelected     (int sel);
-    QGCMAVLinkMessage*  findMessage     (uint32_t id, uint8_t compId);
-    int                 findMessage     (QGCMAVLinkMessage* message);
-    void                append          (QGCMAVLinkMessage* message);
-    QGCMAVLinkMessage*  selectedMsg     ();
-
-signals:
-    void compIDsChanged                 ();
-    void selectedChanged                ();
-
-private:
-    void _checkCompID                   (QGCMAVLinkMessage *message);
-    void _resetSelection                ();
-
-private:
-    quint8              _id;
-    QList<int>          _compIDs;
-    QStringList         _compIDsStr;
-    QmlObjectListModel  _messages;      //-- List of QGCMAVLinkMessage
-    int                 _selected = 0;
-};
-
-//-----------------------------------------------------------------------------
-/// MAVLink message charting controller
-class MAVLinkChartController : public QObject {
-    Q_OBJECT
-public:
-    MAVLinkChartController(MAVLinkInspectorController* parent, int index);
-
-    Q_PROPERTY(QVariantList chartFields         READ chartFields            NOTIFY chartFieldsChanged)
-    Q_PROPERTY(QDateTime    rangeXMin           READ rangeXMin              NOTIFY rangeXMinChanged)
-    Q_PROPERTY(QDateTime    rangeXMax           READ rangeXMax              NOTIFY rangeXMaxChanged)
-    Q_PROPERTY(qreal        rangeYMin           READ rangeYMin              NOTIFY rangeYMinChanged)
-    Q_PROPERTY(qreal        rangeYMax           READ rangeYMax              NOTIFY rangeYMaxChanged)
-    Q_PROPERTY(int          chartIndex          READ chartIndex             CONSTANT)
-
-    Q_PROPERTY(quint32      rangeYIndex         READ rangeYIndex            WRITE setRangeYIndex    NOTIFY rangeYIndexChanged)
-    Q_PROPERTY(quint32      rangeXIndex         READ rangeXIndex            WRITE setRangeXIndex    NOTIFY rangeXIndexChanged)
-
-    Q_INVOKABLE void        addSeries           (QGCMAVLinkMessageField* field, QAbstractSeries* series);
-    Q_INVOKABLE void        delSeries           (QGCMAVLinkMessageField* field);
-
-    Q_INVOKABLE MAVLinkInspectorController* controller() { return _controller; }
-
-    QVariantList            chartFields         () { return _chartFields; }
-    QDateTime               rangeXMin           () { return _rangeXMin;   }
-    QDateTime               rangeXMax           () { return _rangeXMax;   }
-    qreal                   rangeYMin           () const{ return _rangeYMin;   }
-    qreal                   rangeYMax           () const{ return _rangeYMax;   }
-    quint32                 rangeXIndex         () const{ return _rangeXIndex; }
-    quint32                 rangeYIndex         () const{ return _rangeYIndex; }
-    int                     chartIndex          () const{ return _index; }
-
-    void                    setRangeXIndex      (quint32 t);
-    void                    setRangeYIndex      (quint32 r);
-    void                    updateXRange        ();
-    void                    updateYRange        ();
-
-signals:
-    void chartFieldsChanged ();
-    void rangeXMinChanged   ();
-    void rangeXMaxChanged   ();
-    void rangeYMinChanged   ();
-    void rangeYMaxChanged   ();
-    void rangeYIndexChanged ();
-    void rangeXIndexChanged ();
-
-private slots:
-    void _refreshSeries     ();
-
-private:
-    QTimer              _updateSeriesTimer;
-    QDateTime           _rangeXMin;
-    QDateTime           _rangeXMax;
-    int                 _index               = 0;
-    qreal               _rangeYMin           = 0;
-    qreal               _rangeYMax           = 1;
-    quint32             _rangeXIndex         = 0;                    ///< 5 Seconds
-    quint32             _rangeYIndex         = 0;                    ///< Auto Range
-    QVariantList        _chartFields;
-    MAVLinkInspectorController* _controller  = nullptr;
-};
 
 //-----------------------------------------------------------------------------
 /// MAVLink message inspector controller (provides the logic for UI display)
