+++ src/Camera/VehicleCameraControl.cc
@@ -17,6 +17,7 @@
 #include "QGCCorePlugin.h"
 #include "Vehicle.h"
 #include "LinkInterface.h"
+#include "MAVLinkProtocol.h"
 
 #include <QtNetwork/QNetworkAccessManager>
 #include <QtCore/QDir>
