+++ src/GPS/RTCMMavlink.cc
@@ -10,6 +10,7 @@
 
 #include "RTCMMavlink.h"
 #include "MultiVehicleManager.h"
+#include "MAVLinkProtocol.h"
 #include "Vehicle.h"
 
 RTCMMavlink::RTCMMavlink(QGCToolbox& toolbox)
