+++ src/MAVLink/ImageProtocolManager.cc
