+++ src/MAVLink/SysStatusSensorInfo.h
@@ -9,7 +9,7 @@
 
 #pragma once
 
-#include "QGCMAVLink.h"
+#include "MAVLinkLib.h"
 
 #include <QtCore/QMap>
 #include <QtCore/QObject>
