+++ src/Utilities/QGCLoggingCategory.cc
@@ -23,7 +23,6 @@ QGC_LOGGING_CATEGORY(FirmwareUpgradeVerboseLog,     "FirmwareUpgradeVerboseLog")
 QGC_LOGGING_CATEGORY(MissionCommandsLog,            "MissionCommandsLog")
 QGC_LOGGING_CATEGORY(MissionItemLog,                "MissionItemLog")
 QGC_LOGGING_CATEGORY(ParameterManagerLog,           "ParameterManagerLog")
-QGC_LOGGING_CATEGORY(GeotaggingLog,                 "GeotaggingLog")
 QGC_LOGGING_CATEGORY(RTKGPSLog,                     "RTKGPSLog")
 QGC_LOGGING_CATEGORY(GuidedActionsControllerLog,    "GuidedActionsControllerLog")
 QGC_LOGGING_CATEGORY(LocalizationLog,               "LocalizationLog")
