+++ src/Vehicle/InitialConnectStateMachine.h
@@ -17,8 +17,6 @@
 
 Q_DECLARE_LOGGING_CATEGORY(InitialConnectStateMachineLog)
 
-class Vehicle;
-
 class InitialConnectStateMachine : public StateMachine
 {
     Q_OBJECT
