+++ test/MAVLink/CMakeLists.txt
@@ -1,0 +1,3 @@
+find_package(Qt6 REQUIRED COMPONENTS Core)
+
+qt_add_library(MAVLinkTest STATIC)
